requirements.txt
amqp==2.6.1
atomicwrites==1.4.1
attrs==21.4.0
backports-abc==0.5
backports.functools-lru-cache==1.6.4
billiard==3.6.4.0
celery==4.4.7
certifi==2021.10.8
chardet==4.0.0
colorama==0.4.6
configparser==4.0.2
contextlib2==0.6.0.post1
django-redis==4.10.0
Django==1.11.29
enum34==1.1.10
factory-boy==2.12.0
Factory==1.2
Faker==3.0.1
flake8-builtins==1.5.3
flake8-commas==2.1.0
flake8-import-order==0.18.1
flake8-quotes==3.3.0
flake8-tidy-imports==1.1.0
flake8==3.9.2
flower==0.9.5
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.4.0
humanize==1.0.0
idna==2.10
importlib-metadata==2.1.3
inflection==0.3.1
ipaddress==1.0.23
kombu==4.6.11
mccabe==0.6.1
more-itertools==5.0.0
packaging==20.9
pathlib2==2.3.7.post1
pip==20.3.4
pluggy==0.13.1
prometheus-client==0.8.0
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest-django==3.10.0
pytest-factoryboy==2.0.3
pytest==4.6.11
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2023.3
redis==3.5.3
requests==2.27.1
scandir==1.10.0
setuptools==44.1.1
singledispatch==3.7.0
six==1.16.0
text-unidecode==1.3
tornado==5.1.1
typing==3.10.0.0
urllib3==1.26.16
vine==1.3.0
wcwidth==0.2.6
wheel==0.37.1
zipp==1.2.0
